requirements.txt
Flask
Flask-Session
azure-storage-blob
azure-search-documents
python-dotenv
PyPDF2
